+++ src/contracts/MslV4.ts
@@ -1,11 +1,19 @@
 import { Address, Hash } from "viem";
 
-import { filterLogs, LoanV4, OfferV4, RenegotiationV4, Wallet } from '@/blockchain';
+import {
+  filterLogs,
+  LoanV4,
+  LoanV4V5,
+  OfferV4,
+  RenegotiationV4,
+  Wallet,
+} from "@/blockchain";
 import { getContracts } from "@/deploys";
 import { multiSourceLoanABI as multiSourceLoanABIV4 } from "@/generated/blockchain/v4";
 import { getDomain } from "@/utils";
 
 import { Contract } from "./Contract";
+import { ContractMethodReturnType } from "./types";
 
 export class MslV4 extends Contract<typeof multiSourceLoanABIV4> {
   constructor({ walletClient }: { walletClient: Wallet }) {
@@ -209,15 +217,17 @@ export class MslV4 extends Contract<typeof multiSourceLoanABIV4> {
   async repayLoan({
     loan,
     nftReceiver,
+    loanId,
   }: {
     loan: LoanV4;
+    loanId: bigint;
     nftReceiver?: Address;
   }) {
     const receiver = nftReceiver ?? this.wallet.account.address;
 
     const txHash = await this.safeContractWrite.repayLoan([
       receiver,
-      loan.source[0].loanId,
+      loanId,
       loan,
       false,
     ]);
@@ -287,10 +297,7 @@ export class MslV4 extends Contract<typeof multiSourceLoanABIV4> {
     offer: RenegotiationV4;
     loan: LoanV4;
   }) {
-    const txHash = await this.safeContractWrite.refinancePartial([
-      offer,
-      loan,
-    ]);
+    const txHash = await this.safeContractWrite.refinancePartial([offer, loan]);
 
     return {
       txHash,
@@ -317,11 +324,15 @@ export class MslV4 extends Contract<typeof multiSourceLoanABIV4> {
     };
   }
 
-  async liquidateLoan({ loan }: { loan: LoanV4 }) {
-    const txHash = await this.safeContractWrite.liquidateLoan([
-      loan.source[0].loanId,
-      loan,
-    ]);
+  async extendLoan(): ContractMethodReturnType<{
+    loan: LoanV4V5 & { id: string; contractAddress: Address };
+    newLoanId: bigint;
+  }> {
+    throw new Error("Not implemented for V1");
+  }
+
+  async liquidateLoan({ loan, loanId }: { loan: LoanV4; loanId: bigint }) {
+    const txHash = await this.safeContractWrite.liquidateLoan([loanId, loan]);
 
     return {
       txHash,