+++ src/contracts/types.ts
@@ -1,0 +1,6 @@
+import { Hash, TransactionReceipt } from "viem";
+
+export type ContractMethodReturnType<T> = Promise<{
+  txHash: Hash;
+  waitTxInBlock: () => Promise<TransactionReceipt & T>;
+}>;